Other: Extract `createDropdownView()` to own file.
